+++ src/emulator/controller.ts
@@ -42,12 +42,11 @@ import { fileExistsSync } from "../fsutils";
 import { StorageEmulator } from "./storage";
 import { getDefaultDatabaseInstance } from "../getDefaultDatabaseInstance";
 import { getProjectDefaultAccount } from "../auth";
+import { Options } from "../options";
+import { ParsedTriggerDefinition } from "./functionsEmulatorShared";
 
-async function getAndCheckAddress(emulator: Emulators, options: any): Promise<Address> {
-  let host = Constants.normalizeHost(
-    options.config.get(Constants.getHostKey(emulator), Constants.getDefaultHost(emulator))
-  );
-
+async function getAndCheckAddress(emulator: Emulators, options: Options): Promise<Address> {
+  let host = options.config.src.emulators?.[emulator]?.host || Constants.getDefaultHost(emulator);
   if (host === "localhost" && utils.isRunningInWSL()) {
     // HACK(https://github.com/firebase/firebase-tools-ui/issues/332): Use IPv4
     // 127.0.0.1 instead of localhost. This, combined with the hack in
@@ -58,7 +57,7 @@ async function getAndCheckAddress(emulator: Emulators, options: any): Promise<Ad
     host = "127.0.0.1";
   }
 
-  const portVal = options.config.get(Constants.getPortKey(emulator), undefined);
+  const portVal = options.config.src.emulators?.[emulator]?.port;
   let port;
   let findAvailablePort = false;
   if (portVal) {
@@ -200,7 +199,7 @@ export function filterEmulatorTargets(options: any): Emulators[] {
  * @param options
  * @param name
  */
-export function shouldStart(options: any, name: Emulators): boolean {
+export function shouldStart(options: Options, name: Emulators): boolean {
   if (name === Emulators.HUB) {
     // The hub only starts if we know the project ID.
     return !!options.project;
@@ -213,7 +212,7 @@ export function shouldStart(options: any, name: Emulators): boolean {
       return true;
     }
 
-    if (options.config.get("emulators.ui.enabled") === false) {
+    if (options.config.src.emulators?.ui?.enabled === false) {
       // Allow disabling UI via `{emulators: {"ui": {"enabled": false}}}`.
       // Emulator UI is by default enabled if that option is not specified.
       return false;
@@ -226,11 +225,7 @@ export function shouldStart(options: any, name: Emulators): boolean {
   }
 
   // Don't start the functions emulator if we can't find the source directory
-  if (
-    name === Emulators.FUNCTIONS &&
-    emulatorInTargets &&
-    !options.config.get("functions.source")
-  ) {
+  if (name === Emulators.FUNCTIONS && emulatorInTargets && !options.config.src.functions?.source) {
     EmulatorLogger.forEmulator(Emulators.FUNCTIONS).logLabeled(
       "WARN",
       "functions",
@@ -311,7 +306,7 @@ function findExportMetadata(importPath: string): ExportMetadata | undefined {
   }
 }
 
-export async function startAll(options: any, showUI: boolean = true): Promise<void> {
+export async function startAll(options: Options, showUI: boolean = true): Promise<void> {
   // Emulators config is specified in firebase.json as:
   // "emulators": {
   //   "firestore": {
@@ -390,6 +385,7 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
     version: "unknown",
   };
   if (options.import) {
+    utils.assertIsString(options.import);
     const importDir = path.resolve(options.import);
     const foundMetadata = findExportMetadata(importDir);
     if (foundMetadata) {
@@ -407,9 +403,17 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
     const functionsLogger = EmulatorLogger.forEmulator(Emulators.FUNCTIONS);
     const functionsAddr = await getAndCheckAddress(Emulators.FUNCTIONS, options);
     const projectId = getProjectId(options, false);
+
+    utils.assertDefined(options.config.src.functions);
+    utils.assertDefined(
+      options.config.src.functions.source,
+      "Error: 'functions.source' is not defined"
+    );
+
+    utils.assertIsStringOrUndefined(options.extensionDir);
     const functionsDir = path.join(
       options.extensionDir || options.config.projectDir,
-      options.config.get("functions.source")
+      options.config.src.functions.source
     );
 
     let inspectFunctions: number | undefined;
@@ -438,7 +442,7 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
       );
     }
 
-    const account = getProjectDefaultAccount(options.projectRoot);
+    const account = getProjectDefaultAccount(options.projectRoot as string | null);
     const functionsEmulator = new FunctionsEmulator({
       projectId,
       functionsDir,
@@ -447,9 +451,9 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
       port: functionsAddr.port,
       debugPort: inspectFunctions,
       env: {
-        ...options.extensionEnv,
+        ...(options.extensionEnv as Record<string, string> | undefined),
       },
-      predefinedTriggers: options.extensionTriggers,
+      predefinedTriggers: options.extensionTriggers as ParsedTriggerDefinition[] | undefined,
       nodeMajorVersion: parseRuntimeVersion(
         options.extensionNodeVersion || options.config.get("functions.runtime")
       ),
@@ -469,6 +473,7 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
     };
 
     if (exportMetadata.firestore) {
+      utils.assertIsString(options.import);
       const importDirAbsPath = path.resolve(options.import);
       const exportMetadataFilePath = path.resolve(
         importDirAbsPath,
@@ -483,8 +488,8 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
       args.seed_from_export = exportMetadataFilePath;
     }
 
-    const config = options.config as Config;
-    const rulesLocalPath = config.get("firestore.rules");
+    const config = options.config;
+    const rulesLocalPath = config.src.firestore?.rules;
     let rulesFileFound = false;
     if (rulesLocalPath) {
       const rules: string = config.path(rulesLocalPath);
@@ -575,6 +580,7 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
     await startEmulator(databaseEmulator);
 
     if (exportMetadata.database) {
+      utils.assertIsString(options.import);
       const importDirAbsPath = path.resolve(options.import);
       const databaseExportDir = path.resolve(importDirAbsPath, exportMetadata.database.path);
 
@@ -605,6 +611,7 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
     await startEmulator(authEmulator);
 
     if (exportMetadata.auth) {
+      utils.assertIsString(options.import);
       const importDirAbsPath = path.resolve(options.import);
       const authExportDir = path.resolve(importDirAbsPath, exportMetadata.auth.path);
 
@@ -648,6 +655,7 @@ export async function startAll(options: any, showUI: boolean = true): Promise<vo
     await startEmulator(storageEmulator);
 
     if (exportMetadata.storage) {
+      utils.assertIsString(options.import);
       const importDirAbsPath = path.resolve(options.import);
       const storageExportDir = path.resolve(importDirAbsPath, exportMetadata.storage.path);
       storageEmulator.storageLayer.import(storageExportDir);